fix(APISIMP-PREDICATES-PAGECAP): lower ShapesPredicatesRest pageSize cap from 500 to standard 200 (#2428)
* fix(APISIMP-PREDICATES-PAGECAP): lower ShapesPredicatesRest pageSize cap from 500 to standard 200

GET /v2/shapes/predicates used @Max(500) and "1–500" description with no documented
reason to deviate from the v2 standard @Max(200) cap. Predicate vocabulary entries
are lightweight; same fix shape as APISIMP-TPLREST-TAG-PAGECAP (fire-478).


* ci: retrigger CodeQL (transient aggregator flake on fire-512)

* fix(APISIMP-PREDICATES-PAGECAP): add substrate allowlist to resolve CodeQL SQL-injection false positive

Restrict ?substrate= to the four known values (neo4j, timescaledb, postgres,
garage). Unknown values return 400 before reaching the repository layer, which
terminates CodeQL's taint trace and is also correct defensive input validation.
Adds three unit tests for the 400/200 paths around the new allowlist check.


* ci: retrigger CodeQL (aggregator ran before analysis completed on fire-513)

* fix(APISIMP-PREDICATES-PAGECAP): break CodeQL taint chain via re-derive from trusted constant

Re-derive `cleanSub` from `ALLOWED_SUBSTRATES` via stream+findFirst rather than
using the `sub` variable (which still carries user-input taint in CodeQL's model).
Set.contains() alone does not break the taint chain; pulling the value back from the
trusted constant gives CodeQL a clean flow into repository.countBySubstrate /
repository.findBySubstrate.


* ci: retrigger CodeQL (policy check raced scan completion on fire-514)

* ci: retrigger CodeQL (policy check raced scan completion on fire-515)

* fix(APISIMP-PREDICATES-PAGECAP): prevent CWE-190 int overflow in skip arithmetic

Use long skip = (long) page * pageSize in ShapesPredicatesRest to
prevent integer overflow at large page/pageSize values (CWE-190).
Update PredicateVocabularyRepository.findAll and findBySubstrate to
accept long skip; change ps.setInt(2/3, skip) to ps.setLong for OFFSET.

---------

+++ backend/src/main/java/de/dlr/shepard/v2/shapes/repositories/PredicateVocabularyRepository.java
@@ -130,11 +130,11 @@ public class PredicateVocabularyRepository {
    * @param limit max rows to return (LIMIT)
    * @return matching rows; empty list on error
    */
-  public List<PredicateVocabularyEntryIO> findAll(int skip, int limit) {
+  public List<PredicateVocabularyEntryIO> findAll(long skip, int limit) {
     try (Connection conn = defaultDataSource.getConnection();
          PreparedStatement ps = conn.prepareStatement(SELECT_ALL_PAGED)) {
       ps.setInt(1, limit);
-      ps.setInt(2, skip);
+      ps.setLong(2, skip);
       try (ResultSet rs = ps.executeQuery()) {
         return mapRows(rs);
       }
@@ -153,7 +153,7 @@ public class PredicateVocabularyRepository {
    * @param limit max rows to return (LIMIT)
    * @return matching rows; empty list on error or unknown substrate
    */
-  public List<PredicateVocabularyEntryIO> findBySubstrate(String substrate, int skip, int limit) {
+  public List<PredicateVocabularyEntryIO> findBySubstrate(String substrate, long skip, int limit) {
     if (substrate == null || substrate.isBlank()) {
       return Collections.emptyList();
     }
@@ -161,7 +161,7 @@ public class PredicateVocabularyRepository {
          PreparedStatement ps = conn.prepareStatement(SELECT_BY_SUBSTRATE_PAGED)) {
       ps.setString(1, substrate.trim());
       ps.setInt(2, limit);
-      ps.setInt(3, skip);
+      ps.setLong(3, skip);
       try (ResultSet rs = ps.executeQuery()) {
         return mapRows(rs);
       }
